+++ logdissect/morphers/grep.py
@@ -41,8 +41,13 @@ class MorphModule(OurModule):
         else:
             repattern = re.compile(r".*({}).*".format(args.pattern[0]))
 
-            newdata = data
-            newdata.entries = []
+            # newdata = data
+            # newdata.entries = []
+            newdata = LogData()
+            newdata.source_path = data.source_path
+            newdata.source_file = data.source_file
+            newdata.source_file_mtime = data.source_file_mtime
+            newdata.parser = data.parser
 
             for entry in data.entries:
                 if re.match(repattern, entry.raw_text):