+++ logdissect/morphers/last.py
@@ -60,8 +60,13 @@ class MorphModule(OurModule):
             ourstart = starttime.strftime('%Y%m%d%H%M%S')
             
             # Pull out the specified time period:
-            newdata = data
-            newdata.entries = []
+            # newdata = data
+            # newdata.entries = []
+            newdata = LogData()
+            newdata.source_path = data.source_path
+            newdata.source_file = data.source_file
+            newdata.source_file_mtime = data.source_file_mtime
+            newdata.parser = data.parser
             
             for entry in data.entries:
                 if int(entry.date_stamp) >= int(ourstart):