+++ logdissect/morphers/rgrep.py
@@ -40,8 +40,14 @@ class MorphModule(OurModule):
             return data
         else:
             repattern = re.compile(r".*({}).*".format(args.rpattern[0]))
-            newdata = data
-            newdata.entries = []
+            # newdata = data
+            # newdata.entries = []
+            newdata = LogData()
+            newdata.source_path = data.source_path
+            newdata.source_file = data.source_file
+            newdata.source_file_mtime = data.source_file_mtime
+            newdata.parser = data.parser
+
             for entry in data.entries:
                 if not re.match(repattern, entry.raw_text):
                     newdata.entries.append(entry)